Create Makefile

+++ make
@@ -3,11 +3,33 @@
 # Copyright (C) Andrzej Adamczyk (at https://blackdev.org/). All rights reserved.
 #=================================================================================
 
-# external resources initialization
-#git submodule update --init
+# we use clang, as no cross-compiler needed, include std.h header as default for all
+CC="clang"
+C="${CC} -include ./library/std.h -g"
+LD="ld.lld"
+ASM="nasm"
+ISO="xorriso"
+V="qemu-system-x86_64"
+ARCH="x86-64"
+FLAGS="-march=${ARCH} -mtune=generic -O${OPT} -m64 -ffreestanding -nostdlib -nostartfiles -fno-stack-protector -fno-builtin -mno-red-zone ${DEBUG}"
 
 # for clear view
 clear
+
+# check environment software, required!
+ENV=true
+echo -n "Environment: "
+	type -a ${CC} &> /dev/null || (echo -en "\e[38;5;196m${C}\e[0m" && ENV=false)
+	type -a ${LD} &> /dev/null || (echo -en "\e[38;5;196m${LD}\e[0m" && ENV=false)
+	type -a ${ASM} &> /dev/null || (echo -en "\e[38;5;196m${ASM}\e[0m" && ENV=false)
+	type -a ${ISO} &> /dev/null || (echo -en "\e[38;5;196m${ISO}\e[0m" && ENV=false)
+if [ "${ENV}" = false ]; then echo -e "\n[please install missing software]"; exit 1; else echo -e "\e[38;5;2m\xE2\x9C\x94\e[0m"; fi
+
+# optional
+type -a qemu-system-x86_64 &> /dev/null || echo -e "Optional \e[38;5;11mqemu\e[0m not installed. ISO file will be generated regardless of that."
+
+# external resources initialization
+git submodule update --init
 
 # remove all obsolete files, which could interference
 rm -rf build && mkdir -p build/iso
@@ -23,36 +45,26 @@ cp -rf root build
 OPT="${1}"
 if [ -z "${OPT}" ]; then OPT="fast"; fi
 
-# we use clang, as no cross-compiler needed, include std.h header as default for all
-C="clang -include ./library/std.h -g"
-LD="ld.lld"
-ASM="nasm"
-
-# when looking at "Other Settings" tab
-# https://store.steampowered.com/hwsurvey/ (if someone have something better to lookup, please provide)
-# almost everyone have at least Intel Core i3,5,7 or better
-ARCH="x86-64"
-DEFAULT_FLAGS="-march=${ARCH} -mtune=generic -O${OPT} -m64 -ffreestanding -nostdlib -nostartfiles -fno-stack-protector -fno-builtin -mno-red-zone ${DEBUG}"
-
 # build subroutines required by kernel
 EXT=""
-${ASM} -f elf64 kernel/init/gdt.asm	-o build/gdt.o & EXT="${EXT} build/gdt.o"
-${ASM} -f elf64 kernel/idt.asm		-o build/idt.o & EXT="${EXT} build/idt.o"
-${ASM} -f elf64 kernel/task.asm		-o build/task.o & EXT="${EXT} build/task.o"
-${ASM} -f elf64 kernel/rtc.asm		-o build/rtc.o & EXT="${EXT} build/rtc.o"
-# ${ASM} -f elf64 kernel/hpet.asm		-o build/hpet.o & EXT="${EXT} build/hpet.o"
-${ASM} -f elf64 kernel/syscall.asm	-o build/syscall.o & EXT="${EXT} build/syscall.o"
+${ASM} -f elf64 kernel/init/gdt.asm	-o build/gdt.o		|| exit 1; EXT="${EXT} build/gdt.o"
+${ASM} -f elf64 kernel/idt.asm		-o build/idt.o		|| exit 1; EXT="${EXT} build/idt.o"
+${ASM} -f elf64 kernel/task.asm		-o build/task.o		|| exit 1; EXT="${EXT} build/task.o"
+${ASM} -f elf64 kernel/rtc.asm		-o build/rtc.o		|| exit 1; EXT="${EXT} build/rtc.o"
+# ${ASM} -f elf64 kernel/hpet.asm		-o build/hpet.o		|| exit 1; EXT="${EXT} build/hpet.o"
+${ASM} -f elf64 kernel/syscall.asm	-o build/syscall.o	|| exit 1; EXT="${EXT} build/syscall.o"
 
 # default configuration of clang for kernel making
 if [ ! -z "${2}" ]; then DEBUG="-DDEBUG"; fi
-CFLAGS="${DEFAULT_FLAGS} -mno-mmx -mno-sse -mno-sse2 -mno-3dnow ${DEBUG}"
-CFLAGS_SOFTWARE="${DEFAULT_FLAGS} ${DEBUG}"
+CFLAGS="${FLAGS} -mno-mmx -mno-sse -mno-sse2 -mno-3dnow ${DEBUG}"
+CFLAGS_SOFTWARE="${FLAGS} ${DEBUG}"
 LDFLAGS="-nostdlib -static -no-dynamic-linker"
 
 # build kernel file
 ${C} -c kernel/init.c -o build/kernel.o ${CFLAGS} || exit 1;
 ${LD} ${EXT} build/kernel.o -o build/kernel -T tools/kernel.ld ${LDFLAGS} || exit 1;
 strip -s build/kernel
+echo -e "  Kernel\e[38;5;2m\r\xE2\x9C\x94\e[0m"
 
 # copy kernel file and limine files onto destined iso folder
 gzip -k build/kernel
@@ -60,12 +72,13 @@ cp build/kernel.gz tools/limine.cfg limine/limine-bios.sys limine/limine-bios-cd
 
 #===============================================================================
 
-for submodule in `(cd module && ls *.asm)`; do
+for submodules in `(cd module && ls *.asm)`; do
 	# module name
-	name=${submodule%.*}
+	submodule=${submodules%.*}
 
 	# build
-	${ASM} -f elf64 module/${name}.asm -o build/${name}.ao
+	${ASM} -f elf64 module/${submodule}.asm -o build/${submodule}.ao
+	echo -e "  Submodule ${submodule}.asm\e[38;5;2m\r\xE2\x9C\x94\e[0m"
 done
 
 for modules in `(cd module && ls *.c)`; do
@@ -74,6 +87,7 @@ for modules in `(cd module && ls *.c)`; do
 
 	# build
 	${C} -c -fpic -DMODULE module/${module}.c -o build/${module}.o ${CFLAGS} || exit 1
+	echo -e "  Module ${module}.c\e[38;5;2m\r\xE2\x9C\x94\e[0m"
 
 	# connect with libraries (if necessery)
 	SUB=""
@@ -92,6 +106,7 @@ lib=""	# include list of libraries
 for library in path color elf integer string network input math json font std image interface random rgl terminal; do
 	# build
 	${C} -c -fpic library/${library}.c -o build/${library}.o ${CFLAGS_SOFTWARE} || exit 1
+	echo -e "  Library ${library}.c\e[38;5;2m\r\xE2\x9C\x94\e[0m"
 
 	# convert to shared
 	${C} -shared build/${library}.o -o build/root/system/lib/lib${library}.so ${CFLAGS_SOFTWARE} -Wl,--as-needed,-T./tools/library.ld -L./build/root/system/lib/ ${lib} || exit 1
@@ -111,6 +126,7 @@ for software in `(cd software && ls *.c)`; do
 
 	# build
 	${C} -DSOFTWARE -c software/${name}.c -o build/${name}.o ${CFLAGS_SOFTWARE} || exit 1
+	echo -e "  Software ${name}.c\e[38;5;2m\r\xE2\x9C\x94\e[0m"
 
 	# connect with libraries (if necessery)
 	${LD} --as-needed -L./build/root/system/lib build/${name}.o -o build/root/system/bin/${name} ${lib} -T tools/software.ld ${LDFLAGS}
